Refactor log parser
