+++ core/package.json
@@ -3,11 +3,16 @@
   "version": "1.0.0",
   "type": "module",
   "exports": {
+    "./log-parser": "./log-parser/index.js",
     "./constants": "./constants.js",
     "./logger": "./logger.js",
     "./rcon": "./rcon.js"
   },
   "dependencies": {
-    "chalk": "^4.1.0"
+    "async": "^3.2.0",
+    "chalk": "^4.1.0",
+    "ftp-tail": "^1.1.1",
+    "moment": "^2.29.1",
+    "tail": "^2.0.4"
   }
 }
