--- log-parser/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "log-parser",
-  "version": "1.0.0",
-  "type": "module",
-  "exports": {
-    ".": "./index.js"
-  },
-  "dependencies": {
-    "async": "^3.2.0",
-    "ftp-tail": "^1.1.1",
-    "moment": "^2.29.0",
-    "tail": "^2.0.4"
-  }
-}
