+++ squad-server/factory.js
@@ -116,14 +116,19 @@ export default class SquadServerFactory {
     if (type === 'sequelize') {
       let connector;
 
-      if(typeof connectorConfig === 'string') {
-        connector = new Sequelize(connectorConfig, { logging: msg => Logger.verbose('Sequelize', 3, msg) })
+      if (typeof connectorConfig === 'string') {
+        connector = new Sequelize(connectorConfig, {
+          logging: (msg) => Logger.verbose('Sequelize', 3, msg)
+        });
       } else if (typeof connectorConfig === 'object') {
-        connector = new Sequelize({ ...connectorConfig, logging: msg => Logger.verbose('Sequelize', 3, msg) });
+        connector = new Sequelize({
+          ...connectorConfig,
+          logging: (msg) => Logger.verbose('Sequelize', 3, msg)
+        });
       } else {
         throw new Error('Unknown sequelize connector config type.');
       }
-      
+
       await connector.authenticate();
       return connector;
     }
