+++ squad-server/log-parser/new-game.js
