+++ squad-server/log-parser/player-damaged.js
