+++ squad-server/log-parser/player-died.js
