+++ squad-server/log-parser/player-wounded.js
