+++ squad-server/log-parser/steamid-connected.js
