+++ squad-server/package.json
@@ -8,7 +8,6 @@
     "didyoumean": "^1.2.1",
     "discord.js": "^12.3.1",
     "gamedig": "^2.0.20",
-    "log-parser": "1.0.0",
     "mariadb": "^2.5.1",
     "mysql2": "^2.2.5",
     "pg": "^8.5.1",
